🚚 [Move]: #19 object-scanner テストをカテゴリごとに 4 ファイルに分割
レビュー指摘 (#105 thread M) 対応。`object-scanner.behavior.test.ts`
1 ファイル (155 行・23 件) に集約していたテストを `{module}.{aspect}.test.ts`
命名規則に沿って 4 ファイルへ分割する:

- detection: 基本検出（単一/複数/空白バリエーション）
- boundary: token-boundary / 部分一致 / 空入力 / 負号
- skip: ObjectNumber / GenerationNumber 検証で skip される候補
- comment: コメント領域の除外と挟み込みケース
